Fix the unconditional output setting for the router's subflow
* Add `IntegrationFlowDefinition.isOutputChannelRequired()` to check if sub-flow requires an output to the main flow.
* Add `.bridge()` from sub-flows to main flow only if one of sub-flow requires that
* Fix property population on potential Proxy in the `IntegrationFlowDefinition.registerOutputChannelIfCan()`
* Extract `RouterTests` from the `IntegrationFlowTests` to simplify the life during code scanning
* Add test-case when sub-flow doesn't return to main flow
* Validate the continuation of the main flow is mapped to the `defaultOutputChannel`,
if sub-flow doesn't return to main flow
